+++ src/main/java/com/taller/bookstore/dto/response/BookResponse.java
@@ -1,7 +1,7 @@
 package com.taller.bookstore.dto.response;
 
 import lombok.*;
-
+import java.math.BigDecimal;
 import java.util.List;
 
 @Getter
@@ -13,12 +13,11 @@ public class BookResponse {
     private Long id;
     private String title;
     private String isbn;
-    private Double price;
+
+    private BigDecimal price; // ✅ CORREGIDO
+
     private Integer stock;
 
-    // 🔹 NO DEVOLVEMOS ENTIDAD COMPLETA
     private AuthorResponse author;
-
-    // 🔹 EVITA RECURSIÓN
     private List<CategoryResponse> categories;
 }
